+++ src/script-setup/scope-analyzer.ts
